requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-experimental==0.3.4
langchain-openai==0.3.11
langgraph==0.3.21
PyYAML==6.0.2
streamlit==1.44.0
ipython==9.0.2
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.1
tavily-python==0.7.10